+++ Item/Item.js
@@ -93,6 +93,14 @@ const ItemBase = kind({
 
 	propTypes: /** @lends sandstone/Item.ItemBase.prototype */ {
 		/**
+		 * Centers the slots and content.
+		 *
+		 * @type {Boolean}
+		 * @public
+		 */
+		centered: PropTypes.bool,
+
+		/**
 		 * Called with a reference to the root component.
 		 *
 		 * @type {Object|Function}
@@ -202,12 +210,12 @@ const ItemBase = kind({
 		className: ({label, selected, styler}) => styler.append({selected, hasLabel: Boolean(label)})
 	},
 
-	render: ({children, componentRef, css, inline, label, labelPosition, marqueeOn, slotAfter, slotBefore, ...rest}) => {
+	render: ({centered, children, componentRef, css, inline, label, labelPosition, marqueeOn, slotAfter, slotBefore, ...rest}) => {
 		return (
 			<UiItemBase
 				data-webos-voice-intent="Select"
 				component={Row}
-				align="center"
+				align={centered ? 'center center' : 'center'}
 				ref={componentRef}
 				{...rest}
 				inline={inline}
@@ -224,7 +232,7 @@ const ItemBase = kind({
 					label={label}
 					labelPosition={labelPosition}
 					marqueeOn={marqueeOn}
-					shrink={inline}
+					shrink={inline || centered}
 				/>
 				{slotAfter ? (
 					<Cell className={css.slotAfter} shrink>